Extract plotting module from the monolith (split PR 5/N)
The biggest single-concern extraction: moves plot_promoter and
plot_promoter_all — 670+ lines of matplotlib figure assembly — into
tfbs_footprinter3/plotting.py.

- tfbs_footprinter3/plotting.py (new): plot_promoter, plot_promoter_all
- matplotlib.use('Agg') is now set inside plotting.py at its import
  time; callers no longer need to coordinate backend setup before
  importing pyplot. The monolith still calls matplotlib.use('Agg')
  defensively for other (non-plotting) imports, which is harmless.
- The ##-prefixed prior-version alternate signature that preceded
  plot_promoter (a shadow signature including a cage_correlations_hit_tf_dict
  parameter that the active function no longer takes) is dropped as
  behavior-neutral dead-code cleanup.
- tfbs_footprinter3.py re-exports both plot functions so existing
  callers (main(), etc.) continue to work via the stable import path.

Impact:
- Monolith: 1937 -> 1271 lines (-666)
- All 21 unit tests green
- ruff check clean
- Smoke verifies plot_promoter and plot_promoter_all resolve via the
  tfbs_footprinter3.tfbs_footprinter3 path

Remaining in the monolith (~1271 lines):
- ensemblrest, ensemblrest_rate (needs hpc/ensembl_cache.py coordination)
- experimentalDataUpdater, experimentaldata, species_specific_data
- tfbs_finder, alignment_tools, retrieve_genome_aligned
- transfabulator, test_transcript_id, transcript_data_retrieve,
  gene_data_retrieve, retrieve_regulatory
- main, get_args (cli)
